Fix regression introduced in hg revision 25c27d429590.
This commit adds a fix (and regression test) for the following:

Commenting out the block at line 422 of src/indent.c causes indent
to swallow newlines following a rbrace.

To fix this issue, I rewrote the block in question to check more
thoroughly, and further still to check that there's nothing in the
comment buffer before dumping the line. This preserves compatibility
with the 'fix' introduced by the commit in question.