Fix formatting of struct initializer lists, and add -sar
Initializer lists weren't being properly detected as block
initializers and thus were being formatted with a newline forced after
the opening brace.

The newly added -sar option will add spaces between the braces and the
inner part of the initializer for those who prefer it. I've alwo added
this to the K&R style, as this style can be clearly seen in The C
Programming Language 2nd Ed.