Correctly handle pointer alignment with typedef'd types
This fixes a as-of-yet unreported bug, in which indent misaligns pointers
in function parameters when the parameters use a typedef type. For example,

    func f1(int *void)

renders correctly, but

    func f1(uint64_t *void)

misrenders as:

    func f1(uint64_t * void)

The problem occurs because the parser thinks `uint64_t* void` is a binary
multiplication operation, while `int *void` involves a unary dereference
operator. The pointer alignment code only adjust the spacing of unary
operators, at present.

This change fixes the issue by looking for multiplications inside
function parameters and spacing them properly.